requirements.txt
pyserial >= 3.5
prettytable >= 2.1.0